fixup! Add "Create Documents" SearchKit action
