+++ Civi/Civioffice/EventSubscriber/AbstractCoreEntityCiviOfficeTokenSubscriber.php
@@ -17,7 +17,7 @@
 
 declare(strict_types = 1);
 
-namespace Civi\Civioffice\EntitySubscriber;
+namespace Civi\Civioffice\EventSubscriber;
 
 use Civi\Core\Event\GenericHookEvent;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
